requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
colorama==0.4.3
Django==2.2.7
django-cleanup==4.0.0
djangorestframework==3.10.3
more-itertools==8.0.2
mysql-connector-python==8.0.18
packaging==20.0
Pillow==6.2.1
pluggy==0.13.1
protobuf==3.10.0
py==1.8.1
pyparsing==2.4.6
pytest==5.3.2
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
wcwidth==0.1.8
